feat(dashboard-interactivity):implemented basic layer of pub sub as per new design
rename file

UIC-1319
